fix: reject malformed batch response buffers in all fromBuffer() factories (#113)
All six batch fromBuffer() factories now validate that the provided
buffer length is an exact multiple of the expected struct size before
computing the record count. Malformed buffers (partial records) now
throw \InvalidArgumentException instead of silently creating corrupted
batches.

Affected batch types:
- CreateAccountResultBatch (16B struct)
- CreateTransferResultBatch (16B struct)
- AccountBatch (128B struct)
- TransferBatch (128B struct)
- AccountBalanceBatch (128B struct)
- QueryFilterBatch (64B struct)

Empty buffers (0 bytes) remain valid, producing a batch with zero
records.

+++ src/Batch/AccountBalanceBatch.php
@@ -36,7 +36,7 @@ class AccountBalanceBatch extends AbstractBatch
                 $length,
             ));
         }
-        $count = (int) ($length / $structSize);
+        $count = $length / $structSize;
         $batch = new self($count);
         $batch->buffer = $buffer;
         $batch->length = $count;
