+++ src/Batch/AccountBatch.php
@@ -38,7 +38,7 @@ class AccountBatch extends AbstractBatch
                 $length,
             ));
         }
-        $count = (int) ($length / $structSize);
+        $count = $length / $structSize;
         $batch = new self($count);
         $batch->buffer = $buffer;
         $batch->length = $count;
