+++ src/Batch/CreateTransferResultBatch.php
@@ -37,7 +37,7 @@ class CreateTransferResultBatch extends AbstractBatch
                 $length,
             ));
         }
-        $count = (int) ($length / $structSize);
+        $count = $length / $structSize;
         $batch = new self($count);
         $batch->buffer = $buffer;
         $batch->length = $count;
