+++ src/Batch/QueryFilterBatch.php
@@ -35,7 +35,7 @@ class QueryFilterBatch extends AbstractBatch
                 $length,
             ));
         }
-        $count = (int) ($length / $structSize);
+        $count = $length / $structSize;
         $batch = new self($count);
         $batch->buffer = $buffer;
         $batch->length = $count;
